pyproject.toml
[tool.poetry]
name = "elevate-be"
version = "0.1.0"
description = "Elevate"
authors = ["ArJSarmiento <rneljan@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12, <3.13"
aws-lambda-powertools = "^2.40.1"
boto3-stubs = {extras = ["bedrock-agent-runtime"], version = "^1.37.15"}
pynamodb = "6.0.1"
pydantic = "2.10.6"
pytz = "2025.1"

[tool.poetry.group.rag_api.dependencies]
gql = "3.5.2"
aiohttp = "3.11.14"

[tool.poetry.group.suggestions.dependencies]
instructor = "1.7.7"
anthropic = {version = "0.42.0", extras = ["bedrock"]}

[tool.poetry.group.dev.dependencies]
aws-cdk-lib = "2.185.0"
constructs = "^10.4.2"
aws-cdk-aws-lambda-python-alpha = "2.185.0a0"
python-dotenv = "1.0.1"
pre-commit = "^4.1.0"
detect-secrets = "1.5.0"
boto3 = "1.35.76"

[tool.poetry.group.format.dependencies]
pytest = "6.2.5"
ruff = "0.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.format]
quote-style = "single"
docstring-code-line-length = "dynamic"
indent-style = "space"
docstring-code-format = false
line-ending = "auto"


[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = []
